refactor server structure for serverless deployment and add health check endpoint
