+++ api/index.js
@@ -1,0 +1,7 @@
+'use strict';
+const serverless = require('serverless-http');
+const app = require('../app');
+
+module.exports = serverless(app);
+
+
