+++ package.json
@@ -20,7 +20,8 @@
     "jsonwebtoken": "^9.0.2",
     "lowdb": "^3.0.0",
     "mongoose": "^8.18.2",
-    "morgan": "^1.10.1"
+    "morgan": "^1.10.1",
+    "serverless-http": "^3.2.0"
   },
   "devDependencies": {
     "nodemon": "^3.1.10"
